requirements.txt
discord
python-dotenv
steamship